# laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "learn Laravel angular project",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "/api/documentation"
        }
    ],
    "paths": {
...
        "/bands": {
            "get": {
                "tags": [
                    "Bands"
                ],
                "summary": "List Bands",
                "description": "show all bands",
                "operationId": "2d88f149248e1048071c7117b2c78bfb",
                "responses": {
                    "200": {
                        "description": "Hiện danh sách các bands"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
...
        "/bands/store": {
            "post": {
                "tags": [
                    "Bands"
                ],
                "summary": "Create Band",
                "description": "lưu vào db bands mới. chưa có code",
                "operationId": "e2ca1a4eaa932019fa58b3e3b4076d69",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success: A Newly Created Band"
                    },
                    "422": {
                        "description": "Missing mandatory field"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/bands/{id}": {
            "get": {
                "tags": [
                    "Bands"
                ],
                "summary": "get band detail",
                "description": "view thông tin bands qua id. chưa có code",
                "operationId": "cbf44e868e6bb76099cb1a519e45d950",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id that need to be fetched",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid Band id supplied"
                    },
                    "404": {
                        "description": "Band not found"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Bands"
                ],
                "summary": "update band",
                "description": "lưu vào db thông tin sửa bands. chưa có code",
                "operationId": "8678223e2b7d08e0d78b54482634f27c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id that need to be fetched",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success: Return the Band updated"
                    },
                    "422": {
                        "description": "Missing mandatory field"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Bands"
                ],
                "summary": "delete band",
                "description": "xóa khỏi db thông tin bands. chưa có code",
                "operationId": "db8c94d7e25d697de386ab1994687962",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "band id to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Not found"
                    },
                    "204": {
                        "description": "Success: successful deleted"
                    }
                }
            }
        },
        "/bands/{id}/edit": {
            "get": {
                "tags": [
                    "Bands"
                ],
                "summary": "form sửa bands",
                "description": "form sửa thông tin bands. chưa có code",
                "operationId": "d4ea5d030b8dcc75efd5154bd011a102",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id that need to be fetched",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Form sửa band"
                    },
                    "400": {
                        "description": "Invalid Band id supplied"
                    },
                    "404": {
                        "description": "Band not found"
                    }
                }
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "name": "api_key",
                "in": "header",
                "scheme": "api_key"
            }
        }
    }
}